backend/recognition/requirements.txt
python-socketio[client]==4.3.1
requests==2.22.0
